[IMP]buy view 和 test_buy.py 按功能拆分
