+++ buy/tests/__init__.py
@@ -1,4 +1,7 @@
 # -*- coding: utf-8 -*-
-import test_buy
+import test_buy_order
+import test_buy_receipt
+import test_buy_adjust
+import test_payment
 import test_report
 import test_supplier_statements
