+++ buy/views/vendor_goods_view.xml
@@ -1,0 +1,44 @@
+<?xml version="1.0"?>
+<openerp>
+    <data>
+		<!--继承视图-->
+        <record id="buy_supplier_address_form" model="ir.ui.view">
+            <field name="name">buy.supplier.address.form</field>
+            <field name="model">partner</field>
+            <field name="inherit_id" ref="core.supplier_address_form"/>
+            <field name="arch" type="xml">
+                <field name="main_mobile" position="after">
+                    <field name="goods_ids" colspan="2" nolabel="1" groups='goods.view_cost_groups'>
+                        <tree string="goods" editable="bottom">
+                            <field name="goods_id"/>
+                            <field name="price"/>
+                            <field name="code"/>
+                            <field name="name"/>
+                            <field name="min_qty"/>
+                        </tree>
+                    </field>
+                </field>
+            </field>
+        </record>
+
+        <record id="buy_goods_form" model="ir.ui.view">
+            <field name="name">buy.goods.form</field>
+            <field name="model">goods</field>
+            <field name="inherit_id" ref="goods.goods_form"/>
+            <field name="arch" type="xml">
+            <field name="supplier_id" position="after">
+                <field name="vendor_ids" colspan="2" nolabel="1" groups='goods.view_cost_groups'>
+                        <tree string="goods" editable="bottom">
+                            <field name="vendor_id"
+                            	context="{'form_view_ref': 'core.supplier_address_form'}"/>
+                            <field name="price"/>
+                            <field name="code"/>
+                            <field name="name"/>
+                            <field name="min_qty"/>
+                        </tree>
+                    </field>
+                </field>
+            </field>
+        </record>
+	</data>
+</openerp>
